Revert "refactor(activerecord): create module files matching Rails Base includes (#235)"
This reverts commit 2dc36c190bcb45467313fd293e94d2f8ae92ae21.
