--- packages/activerecord/src/aggregations.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Aggregations */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Aggregations {}
