--- packages/activerecord/src/attribute-assignment.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::AttributeAssignment */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface AttributeAssignment {}
