--- packages/activerecord/src/attribute-methods.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::AttributeMethods */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface AttributeMethods {}
