--- packages/activerecord/src/attributes.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Attributes */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Attributes {}
