+++ packages/activerecord/src/autosave.test.ts
@@ -6,11 +6,7 @@ import { describe, it, expect, beforeEach } from "vitest";
 import { Base, registerModel } from "./index.js";
 import { createTestAdapter } from "./test-adapter.js";
 import type { DatabaseAdapter } from "./adapter.js";
-import {
-  markForDestruction,
-  isMarkedForDestruction,
-  isDestroyable,
-} from "./autosave-association.js";
+import { markForDestruction, isMarkedForDestruction, isDestroyable } from "./autosave.js";
 
 function freshAdapter(): DatabaseAdapter {
   return createTestAdapter();
