--- packages/activerecord/src/callbacks.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Callbacks */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Callbacks {}
