--- packages/activerecord/src/connection-handling.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::ConnectionHandling */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface ConnectionHandling {}
