--- packages/activerecord/src/core.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Core */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Core {}
