--- packages/activerecord/src/counter-cache.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::CounterCache */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface CounterCache {}
