--- packages/activerecord/src/dynamic-matchers.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::DynamicMatchers */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface DynamicMatchers {}
