--- packages/activerecord/src/explain.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Explain */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Explain {}
