--- packages/activerecord/src/inheritance.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Inheritance */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Inheritance {}
