--- packages/activerecord/src/integration.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Integration */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Integration {}
