--- packages/activerecord/src/model-schema.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::ModelSchema */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface ModelSchema {}
