--- packages/activerecord/src/no-touching.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::NoTouching */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface NoTouching {}
