--- packages/activerecord/src/normalization.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Normalization */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Normalization {}
