--- packages/activerecord/src/persistence.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Persistence */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Persistence {}
