--- packages/activerecord/src/querying.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Querying */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Querying {}
