--- packages/activerecord/src/readonly-attributes.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::ReadonlyAttributes */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface ReadonlyAttributes {}
