--- packages/activerecord/src/sanitization.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Sanitization */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Sanitization {}
