--- packages/activerecord/src/scoping.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Scoping */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Scoping {}
