--- packages/activerecord/src/serialization.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Serialization */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Serialization {}
