--- packages/activerecord/src/signed-id.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::SignedId */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface SignedId {}
