--- packages/activerecord/src/suppressor.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Suppressor */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Suppressor {}
