--- packages/activerecord/src/timestamp.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Timestamp */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Timestamp {}
