--- packages/activerecord/src/token-for.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::TokenFor */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface TokenFor {}
