--- packages/activerecord/src/touch-later.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::TouchLater */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface TouchLater {}
