+++ packages/activerecord/src/transactions.ts
@@ -90,6 +90,3 @@ export async function savepoint<T>(
     throw error;
   }
 }
-
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Transactions {}
