--- packages/activerecord/src/translation.ts
@@ -1,3 +1,0 @@
-/** Mirrors: ActiveRecord::Translation */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Translation {}
